pyproject.toml
[tool.poetry]
name = "domhunter"
version = "0.1.0"
description = "Domain availability + Google index check + Wayback screenshots pipeline"
authors = ["Your Name <you@example.com>"]
license = "MIT"
readme = "README.md"
packages = [{ include = "domhunter", from = "src" }]

[tool.poetry.dependencies]
python = "^3.10"
httpx = ">=0.27.0"
python-dotenv = ">=1.0.1"

[tool.poetry.group.dev.dependencies]
pytest = "^8.2"
pytest-asyncio = "^0.23"

[tool.poetry.scripts]
domhunter = "domhunter.cli:main"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"